Add files via upload
It deals with choosing between teachers and students, and presents a menu for different functions in that respective case.
The student’s case asks if attendance has to be marked, attendance has to be viewed, or if a student has to view their exam eligibility. For each feature, different functions will be called.
The teacher’s case asks if attendance has to be marked, attendance has to be viewed, work hours has to be viewed, or if a teacher has to generate their pay.